+++ test/snap_reels_test.dart
@@ -4,15 +4,15 @@ import 'package:flutter/material.dart';
 import 'package:snap_reels/snap_reels.dart';
 
 void main() {
-  testWidgets('AwesomeReels widget can be created',
+  testWidgets('SnapReels widget can be created',
       (WidgetTester tester) async {
     await tester.pumpWidget(const MaterialApp(
       home: Scaffold(
-        body: AwesomeReels(
+        body: SnapReels(
           reels: [],
         ),
       ),
     ));
-    expect(find.byType(AwesomeReels), findsOneWidget);
+    expect(find.byType(SnapReels), findsOneWidget);
   });
 }